requirements.txt
# Core Framework
streamlit==1.29.0

# LLM Providers
openai>=1.35.0
groq>=0.9.0
google-generativeai>=0.7.0

# Embeddings and Vector Store
sentence-transformers>=2.3.0
faiss-cpu==1.7.4
huggingface-hub>=0.20.0

# Web Search
duckduckgo-search>=6.1.0

# Utilities
python-dotenv==1.0.0
numpy==1.24.3
pypdf==3.17.1
httpx>=0.27.0

# Optional but recommended
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0